:truck: Rename all references of old repository
