+++ client/ExtensionCompiler.ts
@@ -2,7 +2,7 @@ import { log } from "util";
 import * as webpack from "webpack";
 import ExtensionReloaderImpl from "../src/ExtensionReloader";
 import { info } from "../src/utils/logger";
-import { IPluginOptions } from "../typings/webpack-extension-reloader";
+import { IPluginOptions } from "../typings/webpack-ext-reloader";
 
 export default class ExtensionCompiler {
   private static treatErrors(err) {
