+++ package.json
@@ -3,9 +3,9 @@
   "version": "1.1.8",
   "private": false,
   "description": "Watch for file changes and reload WebExtensions on browsers.",
-  "main": "dist/webpack-extension-reloader.js",
-  "bin": "./dist/webpack-extension-reloader-cli.js",
-  "types": "typings/webpack-extension-reloader.d.ts",
+  "main": "dist/webpack-ext-reloader.js",
+  "bin": "./dist/webpack-ext-reloader-cli.js",
+  "types": "typings/webpack-ext-reloader.d.ts",
   "repository": {
     "type": "git",
     "url": "git://github.com/SimplifyJobs/webpack-ext-reloader.git"
@@ -27,8 +27,8 @@
     "test": "NODE_ENV=test webpack && mocha dist/tests.js",
     "analyze": "NODE_ENV=production webpack --env.analyze",
     "start:dev": "NODE_ENV=development webpack --watch",
-    "start:sample:chrome": "NODE_ENV=development webpack --config sample/webpack.plugin.js --watch",
-    "start:sample:firefox": "NODE_ENV=development webpack --config sample/webpack.plugin.js --watch",
+    "start:sample:chrome": "cross-env NODE_ENV=development cross-env TARGET_BROWSER=chrome webpack --config sample/webpack.plugin.js --watch",
+    "start:sample:firefox": "cross-env NODE_ENV=development cross-env TARGET_BROWSER=firefox webpack --config sample/webpack.plugin.js --watch",
     "prepublishOnly": "npm run build",
     "lint": "eslint \"{src,client,typings,specs}/**/*.ts\" --fix --format codeframe",
     "format": "prettier --write \"{src,client,typings,specs}/**/*.ts\"",
@@ -53,7 +53,9 @@
     "@types/webextension-polyfill": "^0.8.2",
     "@types/webpack": "^5.28.0",
     "@types/webpack-sources": "^3.2.0",
+    "clean-webpack-plugin": "^4.0.0",
     "colors": "^1.4.0",
+    "cross-env": "^7.0.3",
     "lodash": "^4.17.21",
     "minimist": "^1.2.5",
     "useragent": "^2.3.0",
